requirements_test.txt
pytest>=7.0.0
pytest-homeassistant-custom-component>=0.13.0
pytest-asyncio>=0.20.0
